chore: update the readme
Update the README.md file to reflect the current state of the project.
This includes adding a section on how to build and contribute to the
project and updates the note about who is currently maintaining the
project.

This change-set also removes the CHANGELOG.md file as most Deno projects
don't include this and it was mostly a change log of the Astro project.